+++ src/main/java/ru/netology/netologyspringboot/advice/ExceptionHandlerAdvice.java
@@ -1,0 +1,24 @@
+package ru.netology.netologyspringboot.advice;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import ru.netology.netologyspringboot.exceptions.InvalidCredentials;
+import ru.netology.netologyspringboot.exceptions.UnauthorizedUser;
+
+@RestControllerAdvice
+public class ExceptionHandlerAdvice {
+
+    @ExceptionHandler(InvalidCredentials.class)
+    public ResponseEntity<String> ice(InvalidCredentials e) {
+        return new ResponseEntity<>(e.getLocalizedMessage(), HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(UnauthorizedUser.class)
+    public ResponseEntity<String> uue(UnauthorizedUser e) {
+        System.out.println(e.getLocalizedMessage());
+        return new ResponseEntity<>(e.getLocalizedMessage(), HttpStatus.UNAUTHORIZED);
+    }
+
+}
