+++ src/main/java/ru/netology/netologyspringboot/exceptions/InvalidCredentials.java
@@ -1,22 +1,8 @@
 package ru.netology.netologyspringboot.exceptions;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
 public class InvalidCredentials extends RuntimeException {
-
     public InvalidCredentials(String msg) {
         super(msg);
     }
 
-    public InvalidCredentials() {
-    }
-
-    @ExceptionHandler(InvalidCredentials.class)
-    public ResponseEntity<String> ice(InvalidCredentials e) {
-        return new ResponseEntity<>(e.getLocalizedMessage(), HttpStatus.BAD_REQUEST);
-    }
 }
