+++ src/main/java/ru/netology/netologyspringboot/exceptions/UnauthorizedUser.java
@@ -1,23 +1,8 @@
 package ru.netology.netologyspringboot.exceptions;
 
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
 public class UnauthorizedUser extends RuntimeException {
     public UnauthorizedUser(String msg) {
         super(msg);
-    }
-
-    public UnauthorizedUser() {
-    }
-
-    @ExceptionHandler(UnauthorizedUser.class)
-    public ResponseEntity<String> uue(UnauthorizedUser e) {
-        return new ResponseEntity<>(e.getLocalizedMessage(), HttpStatus.UNAUTHORIZED);
     }
 
 }
